update 0.01
hitting alot of errors. with generate function, everything works up until line 54 where it enters the final conditional. Sometimes there are errors where probability does not equal 1 in the np.random.choice call, and alot of times it will choose a character, recurse several times but once it hits a 100% chance of newline, it returns none instead of the commit.