update 0.2
it now works. I redesigned the flow of it to use an object for each word. As it is now, it is kind of determinate, as it it will generate words, or rather parts of words directly from the list. More robust functionality will be added in subsequent versions, but the skeleton is there now.